Moved semantic token range request handler to its own file.
