+++ crates/ty_server/src/server/api/requests.rs
@@ -4,6 +4,7 @@ mod goto_type_definition;
 mod hover;
 mod inlay_hints;
 mod semantic_tokens;
+mod semantic_tokens_range;
 mod shutdown;
 
 pub(super) use completion::CompletionRequestHandler;
@@ -11,5 +12,6 @@ pub(super) use diagnostic::DocumentDiagnosticRequestHandler;
 pub(super) use goto_type_definition::GotoTypeDefinitionRequestHandler;
 pub(super) use hover::HoverRequestHandler;
 pub(super) use inlay_hints::InlayHintRequestHandler;
-pub(super) use semantic_tokens::{SemanticTokensRangeRequestHandler, SemanticTokensRequestHandler};
+pub(super) use semantic_tokens::SemanticTokensRequestHandler;
+pub(super) use semantic_tokens_range::SemanticTokensRangeRequestHandler;
 pub(super) use shutdown::ShutdownHandler;
